rename files from snapshot_replacement* -> region_snapshot_replacement*
