+++ nexus/db-model/src/lib.rs
@@ -72,6 +72,8 @@ mod region;
 mod region_replacement;
 mod region_replacement_step;
 mod region_snapshot;
+mod region_snapshot_replacement;
+mod region_snapshot_replacement_step;
 mod role_assignment;
 mod role_builtin;
 pub mod saga_types;
@@ -90,8 +92,6 @@ mod sled_resource_kind;
 mod sled_state;
 mod sled_underlay_subnet_allocation;
 mod snapshot;
-mod snapshot_replacement;
-mod snapshot_replacement_step;
 mod ssh_key;
 mod switch;
 mod tuf_repo;
@@ -176,6 +176,8 @@ pub use region::*;
 pub use region_replacement::*;
 pub use region_replacement_step::*;
 pub use region_snapshot::*;
+pub use region_snapshot_replacement::*;
+pub use region_snapshot_replacement_step::*;
 pub use role_assignment::*;
 pub use role_builtin::*;
 pub use saga_types::*;
@@ -194,8 +196,6 @@ pub use sled_resource_kind::*;
 pub use sled_state::*;
 pub use sled_underlay_subnet_allocation::*;
 pub use snapshot::*;
-pub use snapshot_replacement::*;
-pub use snapshot_replacement_step::*;
 pub use ssh_key::*;
 pub use switch::*;
 pub use switch_interface::*;
