+++ nexus/db-queries/src/db/datastore/mod.rs
@@ -83,6 +83,7 @@ mod rack;
 mod region;
 mod region_replacement;
 mod region_snapshot;
+mod region_snapshot_replacement;
 mod role;
 mod saga;
 mod silo;
@@ -91,7 +92,6 @@ mod silo_user;
 mod sled;
 mod sled_instance;
 mod snapshot;
-mod snapshot_replacement;
 mod ssh_key;
 mod switch;
 mod switch_interface;
